Build fix.

+++ src/mac/common/GMemDC.cpp
@@ -114,7 +114,7 @@ void CGImg::Create(int x, int y, int Bits, ssize_t Line, uchar *data, uchar *pal
 	#endif
 	
 	if (d->Debug)
-		printf("%s:%i - Bytes=%i, Base=%i, Line=%i, pos=%i,%i\n", _FL, Bytes, (int)(Base-data), Line,
+		printf("%s:%i - Bytes=%i, Base=%i, Line=%i, pos=%i,%i\n", _FL, Bytes, (int)(Base-data), (int)Line,
 			B.x1, B.y1);
 	
 	if (Bits <= 8)
